requirements.txt
streamlit
pandas
xlrd
openpyxl
numpy
matplotlib
scipy
plotly
kaleido
Pillow
streamlit-extras
streamlit-analytics
python-docx
st-pages==0.5.0
openai